Fix TS types being initialzied before global declaration js_only
