+++ assets/prototype/application/valueObjects/config/index.ts
@@ -7,3 +7,5 @@ export * from './DateTimeFormats';
 export * from './Locale';
 export * from './SiteUrl';
 export * from './Timezone';
+
+export * from './types';
