+++ assets/prototype/infrastructure/services/apollo/Apollo.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import pick from 'ramda/src/pick';
+import { pathOr } from 'ramda';
 import { ApolloClient } from 'apollo-client';
-import { InMemoryCache, InMemoryCacheConfig, CacheResolver, NormalizedCacheObject } from 'apollo-cache-inmemory';
+import { InMemoryCache, InMemoryCacheConfig, CacheResolver } from 'apollo-cache-inmemory';
 import { HttpLink } from 'apollo-link-http';
 import { ApolloProvider } from '@apollo/react-hooks';
 
-const { graphqlEndpoint } = window;
-const nonce = pick<Window, string>(['eejsdata', 'data', 'eejs_api_nonce'], window);
+const graphqlEndpoint = pathOr<string>('', ['graphqlEndpoint'], window);
+const nonce = pathOr<string>('', ['eejsdata', 'data', 'eejs_api_nonce'], window);
 
 const getResolver = (type: string): CacheResolver => {
 	const resolver: CacheResolver = (_, args, { getCacheKey }) => getCacheKey({ __typename: type, id: args.id });
@@ -33,11 +33,11 @@ const link = new HttpLink({
 	},
 });
 
-const client = new ApolloClient<NormalizedCacheObject>({
+const client = new ApolloClient({
 	cache,
 	link,
 });
 
-const Apollo = ({ children }) => <ApolloProvider client={client}>{children}</ApolloProvider>;
+const Apollo = ({ children }) => <ApolloProvider client={client} children={children} />;
 
 export default Apollo;
